requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.6
cffi==1.12.2
crl.interactivesessions==1.1.4
crl.remotescript==1.0.1
crl.remotesession==1.2.1
crl.rfcli==1.1.1
crl.threadverify==1.0
cryptography==2.6.1
enum34==1.1.6
future==0.17.1
ipaddress==1.0.22
paramiko==2.4.2
pexpect==4.7.0
ptyprocess==0.6.0
pyasn1==0.4.5
pycparser==2.19
PyNaCl==1.3.0
PyYAML==5.1
robotframework==3.1.1
six==1.12.0
